+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [1-alpha12-SNAPSHOT]
+- patched clerk to pass code metadata freely
+
+
 ## [1-alpha11] - 2022-04-24
 - scittle tool: reordered js lib rendering, added visual spacing